# [Internal] Performance: Improve performance of authorization further.