--- a/Microsoft.Azure.Cosmos/src/StringHMACSHA256Hash.cs
+++ b/Microsoft.Azure.Cosmos/src/StringHMACSHA256Hash.cs
@@ -5,6 +5,7 @@
 namespace Microsoft.Azure.Cosmos
 {
     using System;
+    using System.Collections.Concurrent;
     using System.IO;
     using System.Security;
     using System.Security.Cryptography;
@@ -14,18 +15,33 @@ namespace Microsoft.Azure.Cosmos
         private readonly String base64EncodedKey;
         private readonly byte[] keyBytes;
         private SecureString secureString;
+        private ConcurrentQueue<HMACSHA256> hmacPool;
 
         public StringHMACSHA256Hash(String base64EncodedKey)
         {
             this.base64EncodedKey = base64EncodedKey;
             this.keyBytes = Convert.FromBase64String(base64EncodedKey);
+            this.hmacPool = new ConcurrentQueue<HMACSHA256>();
         }
 
         public byte[] ComputeHash(MemoryStream bytesToHash)
         {
-            using (HMACSHA256 hmacSha256 = new HMACSHA256(this.keyBytes))
+            if (this.hmacPool.TryDequeue(out HMACSHA256 hmacSha256))
             {
-                return hmacSha256.ComputeHash(bytesToHash);
+                hmacSha256.Initialize();
+            }
+            else
+            {
+                hmacSha256 = new HMACSHA256(this.keyBytes);
+            }
+
+            try
+            {
+                return hmacSha256.ComputeHash(bytesToHash.GetBuffer(), 0, (int)bytesToHash.Length);
+            }
+            finally
+            {
+                this.hmacPool.Enqueue(hmacSha256);
             }
         }