# Pull Request Template

## Description

- Add a performance benchmark only for the Authorization layer.
- Add pooling for the HMac pool in StringHMacSha256

Here's the perf results:
1) Without the earlier change (to use Span/MemoryStream etc):
|                     Method |     Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------- |---------:|----------:|----------:|-------:|------:|------:|----------:|
| GenerateAuthorizationToken | 5.242 us | 0.0584 us | 0.3031 us | 0.0992 |     - |     - |   1.61 KB |

2) With the earlier change (to use Span/MemoryStream etc):
|                     Method |     Mean |     Error |    StdDev |   Median |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------- |---------:|----------:|----------:|---------:|-------:|------:|------:|----------:|
| GenerateAuthorizationToken | 6.465 us | 0.1005 us | 0.5239 us | 6.383 us | 0.0916 |     - |     - |   1.41 KB |

3) With the HMac Pooling:
|                     Method |     Mean |     Error |    StdDev |   Median |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------- |---------:|----------:|----------:|---------:|-------:|------:|------:|----------:|
| GenerateAuthorizationToken | 4.194 us | 0.0557 us | 0.2828 us | 4.138 us | 0.0687 |     - |     - |   1.09 KB |

## Type of change

Please delete options that are not relevant.

- [x] Bug fix (non-breaking change which fixes an issue)
- [] New feature (non-breaking change which adds functionality)
- [] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [] This change requires a documentation update

## Closing issues

To automatically close an issue: closes #IssueNumber